abstract authentication providers into a separate file
